+++ R/visualization.R
@@ -2940,7 +2940,8 @@ ISpatialFeaturePlot <- function(
 #' themeing will not work when plotting multiple features/groupings
 #' @param pt.size.factor Scale the size of the spots.
 #' @param alpha Controls opacity of spots. Provide as a vector specifying the
-#' min and max
+#' min and max for SpatialFeaturePlot. For SpatialDimPlot, provide a single
+#' alpha value for each plot.
 #' @param stroke Control the width of the border around the spots
 #' @param interactive Launch an interactive SpatialDimPlot or SpatialFeaturePlot
 #' session, see \code{\link{ISpatialDimPlot}} or
@@ -3177,6 +3178,11 @@ SpatialPlot <- function(
         cols = cols,
         alpha.by = if (is.null(x = group.by)) {
           features[j]
+        } else {
+          NULL
+        },
+        pt.alpha = if (!is.null(x = group.by)) {
+          alpha[j]
         } else {
           NULL
         },
@@ -7495,6 +7501,7 @@ SingleRasterMap <- function(
 # By default, ggplot2 assigns colors
 # @param image.alpha Adjust the opacity of the background images. Set to 0 to
 # remove.
+# @param pt.alpha Adjust the opacity of the points if plotting a SpatialDimPlot
 # @param crop Crop the plot in to focus on points plotted. Set to FALSE to show
 # entire background image.
 # @param pt.size.factor Sets the size of the points relative to spot.radius
@@ -7521,6 +7528,7 @@ SingleSpatialPlot <- function(
   image,
   cols = NULL,
   image.alpha = 1,
+  pt.alpha = NULL,
   crop = TRUE,
   pt.size.factor = NULL,
   stroke = 0.25,
@@ -7561,14 +7569,27 @@ SingleSpatialPlot <- function(
   plot <- switch(
     EXPR = geom,
     'spatial' = {
-      plot + geom_spatial(
-        point.size.factor = pt.size.factor,
-        data = data,
-        image = image,
-        image.alpha = image.alpha,
-        crop = crop,
-        stroke = stroke
-      ) + coord_fixed() + theme(aspect.ratio = 1)
+      if (is.null(x = pt.alpha)) {
+        plot <- plot + geom_spatial(
+          point.size.factor = pt.size.factor,
+          data = data,
+          image = image,
+          image.alpha = image.alpha,
+          crop = crop,
+          stroke = stroke,
+        )
+      } else {
+        plot <- plot + geom_spatial(
+          point.size.factor = pt.size.factor,
+          data = data,
+          image = image,
+          image.alpha = image.alpha,
+          crop = crop,
+          stroke = stroke,
+          alpha = pt.alpha
+        )
+      }
+      plot + coord_fixed() + theme(aspect.ratio = 1)
     },
     'interactive' = {
       plot + geom_spatial_interactive(