+++ NEWS.md
@@ -3,10 +3,12 @@
 ## Added
 
 ## Changes
-- Fix issues with `as.SingleCellExperiment.Seurat` for the latest verion of SingleCellExperiment ([#4532](https://github.com/satijalab/seurat/pull/4532))
+- Fix issues with `as.SingleCellExperiment.Seurat()` for the latest verion of SingleCellExperiment ([#4532](https://github.com/satijalab/seurat/pull/4532))
 - Ensure proper reference.reduction is used in `MapQuery()`
 - Fix to `UpdateSymbolList()`, no longer searches aliases and exposes the `search.types` parameter in `GeneSymbolThesarus()` ([#4545](https://github.com/satijalab/seurat/issues/4545))
 - Transfer `scale.data` slot as well when converting with `as.SingleCellExperiment.Seurat()`
+- Enable `alpha` parameter for `SpatialDimPlot()`
+- Fix `as.SingleCellExperiment.Seurat()` conversion for atypical `reducedDim` components
 
 # Seurat 4.0.2 (2020-03-20)
 ## Added